+++ packages/overlayscrollbars-react/test/useOverlayScrollbars.test.tsx
@@ -2,8 +2,8 @@ import { useRef } from 'react';
 import { describe, test, afterEach, expect } from 'vitest';
 import { render, screen, cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { OverlayScrollbars } from 'overlayscrollbars';
 import { useOverlayScrollbars } from '~/overlayscrollbars-react';
-import type { OverlayScrollbars } from 'overlayscrollbars';
 
 describe('useOverlayScrollbars', () => {
   afterEach(() => cleanup());
@@ -13,25 +13,22 @@ describe('useOverlayScrollbars', () => {
       const instanceRef = useRef<OverlayScrollbars | null>(null);
       const [initialize, instance] = useOverlayScrollbars();
       return (
-        <>
-          <button
-            onClick={(event) => {
-              const osInstance = initialize(event.target as HTMLElement);
-              if (instanceRef.current) {
-                expect(instanceRef.current).toBe(osInstance);
-                expect(instanceRef.current).toBe(instance());
-              }
-              instanceRef.current = osInstance;
+        <button
+          onClick={(event) => {
+            initialize(event.target as HTMLElement);
+            if (instanceRef.current) {
               expect(instanceRef.current).toBe(instance());
-            }}
-          >
-            initialize
-          </button>
-        </>
+            }
+            instanceRef.current = instance();
+            expect(instanceRef.current).toBe(instance());
+          }}
+        >
+          initialize
+        </button>
       );
     };
 
-    render(<Test />);
+    const { unmount } = render(<Test />);
 
     const initializeBtn = screen.getByRole('button');
     userEvent.click(initializeBtn);
@@ -41,5 +38,11 @@ describe('useOverlayScrollbars', () => {
     userEvent.click(initializeBtn);
 
     expect(snapshot).toBe(initializeBtn.innerHTML);
+
+    expect(OverlayScrollbars(initializeBtn)).toBeDefined();
+
+    unmount();
+
+    expect(OverlayScrollbars(initializeBtn)).toBeUndefined();
   });
 });